+++ lib/td_dd/data_structures/data_structure_search.ex
@@ -42,6 +42,7 @@ defmodule TdDd.DataStructure.Search do
     filter_clause = create_filters(params)
     query = create_query(%{}, filter_clause)
     search = %{size: 0, query: query, aggs: agg_terms}
+
     search
     |> Search.search()
     |> get_aggregations_results(agg_terms)
@@ -63,20 +64,64 @@ defmodule TdDd.DataStructure.Search do
     filter = permissions |> create_filter_clause(user_defined_filters)
     query = create_query(%{}, filter)
     search = %{size: 0, query: query, aggs: agg_terms}
+
     search
     |> Search.search()
     |> get_aggregations_results(agg_terms)
   end
 
   defp get_aggregations_results(results, agg_terms) do
-    [agg_term] = Map.keys(agg_terms)
-    results = results
-    |> Map.get(:aggregations, %{})
-    |> Map.get(agg_term, %{})
-    |> Map.get("buckets")
-    case results do
-      nil -> []
-      _ -> results
+    agg_names = get_aggregations_names(%{"aggs" => agg_terms})
+    results = results |> Map.get(:aggregations)
+    get_agg_results(agg_names, results)
+  end
+
+  defp get_agg_results([agg_name | agg_names], results) do
+    results =
+      results
+      |> Map.get(agg_name, %{})
+      |> Map.get("buckets", [])
+
+    r =
+      Enum.map(results, fn bucket ->
+        agg_name_values = %{
+          "type" => agg_name,
+          "doc_count" => bucket["doc_count"],
+          "key" => bucket["key"]
+        }
+
+        case agg_names do
+          [] ->
+            agg_name_values
+
+          _ ->
+            Map.put(agg_name_values, "aggs", get_agg_results(agg_names, bucket))
+        end
+      end)
+  end
+
+  @doc """
+  Extracts aggregations name from aggs format like:
+    %{
+    "systems" => %{
+      :terms => %{field: "system.name.raw"},
+      "aggs" => %{
+        "types" => %{
+          :terms => %{field: "type.raw"},
+          "aggs" => %{"groups" => %{terms: %{field: "group.raw"}}}
+        }
+      }
+    }
+  """
+  defp get_aggregations_names(agg_terms) do
+    case Map.get(agg_terms, "aggs") do
+      nil ->
+        []
+
+      aggs ->
+        [agg_key] = Map.keys(aggs)
+        key_aggs = Map.get(aggs, agg_key)
+        [agg_key] ++ get_aggregations_names(key_aggs)
     end
   end