+++ lib/td_dd/search/aggregations.ex
@@ -5,11 +5,9 @@ defmodule TdDd.Search.Aggregations do
   alias TdCache.TemplateCache
   alias TdDfLib.Format
 
-  @systems_aggregation_term "system.name.raw"
-
   def aggregation_terms do
     static_keywords = [
-      {@systems_aggregation_term, %{terms: %{field: "system.name.raw", size: 50}}},
+      {"system.name.raw", %{terms: %{field: "system.name.raw", size: 50}}},
       {"domain.name.raw", %{terms: %{field: "domain.name.raw", size: 50}}},
       {"group.raw", %{terms: %{field: "group.raw", size: 50}}},
       {"type.raw", %{terms: %{field: "type.raw", size: 50}}},
@@ -26,8 +24,25 @@ defmodule TdDd.Search.Aggregations do
     |> Enum.into(%{})
   end
 
-  def get_systems_agg_terms do
-    Map.take(aggregation_terms(), [@systems_aggregation_term])
+  def get_agg_terms([]) do
+    nil
+  end
+
+  def get_agg_terms([agg_definition | agg_defs]) do
+      %{"agg_name" => agg_name, "field_name" => field_name} = agg_definition
+      new_term_value = %{terms: %{field: field_name, size: 50}}
+      new_term_value = case agg_defs do
+        [] ->
+          new_term_value
+        _ ->
+          new_agg_term_aggs = get_agg_terms(agg_defs)
+          Map.put(new_term_value, "aggs", new_agg_term_aggs)
+      end
+      %{agg_name => new_term_value}
+  end
+
+  def get_agg_term(agg_name, field_name) do
+    Map.put(%{}, agg_name, %{terms: %{field: field_name}})
   end
 
   def template_terms(%{content: content}) do